Factor minrnn into mixer
